Globally clear db in test suite

+++ src/__tests__/testdatabase.ts
@@ -8,5 +8,7 @@ if (process.env.NODE_ENV !== 'test') {
 
 export const clearDb = async () => {
   const db = getDb()
-  await db.query(`TRUNCATE TABLE job_run CASCADE`)
+  if (db) {
+    await db.query(`TRUNCATE TABLE job_run CASCADE`)
+  }
 }
